* Engine
  * Resolution changes / Full-screen mode
  * Debug hud improvements
  * Skipping mesh rendering by distance and radius
  * Run-time shader define controls for particle materials
  * Improved trail renderer

* Physics
  * Force compound option for collision shapes
  * Transformatino reflections for attached rigid bodies

* UI
  * Color ramp 
  * Improved the Drop-down component.
  * Fixed a bug on multi-line text
  * Mouse event propagation between parents and children
  * Added KeyValueDataSet for the ListBox component
  * Buttons can trigger mouse hovering events even in the disabled status
  * Added UIProperty : NO_FOCUS_BY_CLICK, SEND_EVENT_TO_CHILDREN, RECEIVE_EVENT_FROM_PARENT

+++ Code/Engine/ICamera.h
@@ -45,8 +45,8 @@ namespace fastbird
 		virtual void SetNearFar(float nearPlane, float farPlane) = 0;
 		virtual void GetNearFar(float& nearPlane, float& farPlane) const = 0;
 		// width and height of the view volume at the near view-plane
-		virtual void SetWidth(int width) = 0;
-		virtual void SetHeight(int height) = 0;
+		virtual void SetWidth(float width) = 0;
+		virtual void SetHeight(float height) = 0;
 		virtual float GetWidth() const = 0;
 		virtual float GetHeight() const = 0;
 		virtual void SetName(const char* name) = 0;
